backend/requirements.txt
flask==3.0.3
python-dotenv==1.0.1

# Azure + OpenAI
openai==1.107.1
azure-identity==1.17.1

# LangChain ecosystem (Pydantic v2)
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.33
langchain-text-splitters==0.3.11   # >=0.3.9 to satisfy langchain 0.3.27

# Vector store + RAG
chromadb==0.5.5
llama-index==0.11.13

# Data handling
pandas==2.2.2
openpyxl==3.1.5

# Bot Framework
botbuilder-core==4.14.8

# Presidio for PII anonymization
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

# NLP model support
spacy==3.7.5
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl